chore: added method to cache friendships metadata
